requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
blinker==1.8.2
category-encoders==2.6.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dash==2.17.0
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
dominate==2.9.1
executing==2.0.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.2.1
fonttools==4.51.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imbalanced-learn==0.12.2
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain-core==0.2.30
langchain-openai==0.1.21
langsmith==0.1.99
lightgbm==4.3.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.40.6
orjson==3.10.7
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
plotly==5.23.0
plotly-resampler==0.10.0
pmdarima==2.0.4
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycaret==3.3.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyod==1.1.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
regex==2024.7.24
requests==2.32.3
scikit-learn==1.5.0
scikit-plot==0.3.7
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sktime==0.26.0
sniffio==1.3.1
stack-data==0.6.3
statsmodels==0.14.2
tbats==1.1.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
visitor==0.1.3
wcwidth==0.2.13
Werkzeug==3.0.3
WTForms==3.1.2
yellowbrick==1.5
